added validation to all tuple inputs for transforms
