+++ frontend/playground-frontend/src/components/helper_functions/TrainButtonFunctions.js
@@ -9,7 +9,7 @@ const tupleRegex = /^\(([1-9]{1}[0-9]*), ?([1-9]{1}[0-9]*)\)$/;
 export const validateParameter = (source, index, parameter) => {
   const { parameter_name, min, max, parameter_type } = parameter;
   let { value } = parameter;
-  if (parameter_name === "(H, W)") {
+  if (parameter_type === "tuple") {
     if (tupleRegex.test(value)) {
       const result = value.match(tupleRegex);
       const H = result[1].valueOf();
@@ -19,14 +19,14 @@ export const validateParameter = (source, index, parameter) => {
         toast.error(
           `${source} Layer ${
             index + 1
-          }: H not an integer in range [${min}, ${max}]`
+          }: X not an integer in range [${min}, ${max}]`
         );
         return false;
       } else if (W < min || W > max) {
         toast.error(
           `${source} Layer ${
             index + 1
-          }: W not an integer in range [${min}, ${max}]`
+          }: Y not an integer in range [${min}, ${max}]`
         );
         return false;
       }
@@ -35,7 +35,7 @@ export const validateParameter = (source, index, parameter) => {
     toast.error(
       `${source} Layer ${
         index + 1
-      }: ${parameter_name} not of appropriate format: (H, W)`
+      }: ${parameter_name} not of appropriate format: (X, Y)`
     );
   } else {
     if (parameter_type !== "number") return true;
